refactor(server): extract dispatcher from app/mod.rs
`pub fn run()` (586 LOC) and its sync helper `run_sync_with_adapters`
(17 LOC) routed parsed CLI commands to handlers in
`crate::commands::*`. They had no business logic — pure dispatch —
and dominated `app/mod.rs` (1359 → 731 LOC, 47% reduction).

Move both to `app/dispatcher.rs`. Public API preserved via
`pub use dispatcher::run` and `pub(crate) use
dispatcher::run_sync_with_adapters`; the test file
`app/tests/sync.rs` continues to find the helper through the
wildcard re-export.

Net: -628 LOC in mod.rs (the imports and helpers go too); +649 LOC
in the new dispatcher.rs (with new module docstring). 2,025
workspace lib tests pass.
